src/nwp_collector/requirements.txt
pandas==1.1.5
numpy==1.19.5
scikit-learn==0.24.1
scipy==1.5.4
pytest==6.2.2
cassandra-driver==3.25.0
python-dotenv
loguru
netCDF4